+++ src/Utils/RateLimitUtils.php
@@ -7,22 +7,22 @@ class RateLimitUtils
 
     public static function getDaysFrom(\DateTime $date): int
     {
-        return floor(self::getHoursFrom($date) / 24);
+        return (int) floor(self::getHoursFrom($date) / 24);
     }
 
     public static function getHoursFrom(\DateTime $date): int
     {
-        return floor(self::getMinutesFrom($date) / 60);
+        return (int) floor(self::getMinutesFrom($date) / 60);
     }
 
     public static function getMinutesFrom(\DateTime $date): int
     {
-        return floor(self::getSecondsFrom($date) / 60);
+        return (int) floor(self::getSecondsFrom($date) / 60);
     }
 
     public static function getSecondsFrom(\DateTime $date): int
     {
-        return abs(time() - $date->getTimestamp());
+        return (int) abs(time() - $date->getTimestamp());
     }
 
 }
